Publish authorised_macs file to S3
This calls the services to generate and upload the configuration file to
S3 based on the contents of the database.
Stubbed out when running tests or locally.

Currently generates and uploads the authorised_macs file after each change.
Whether we will separate deployments from admin updates is an ongoing
discussion and can be implemented in the future.