Allow larger values to be entered in the Policy Engine
Currently the value fields for both request and responses of the
policy engine is defined as type varchar(255).
This restricts the length of the values that can be entered.
There will be scenarios where larger values need to be entered.

Update the schema for these columns to text to allow this.