Remove ERD from Readme
This will get out of date if not manually re-generated.
The schema.rb file should be the source of truth for this.

Remove documentation around rotating RDS CA bundle, we pull this bundle each
time we deploy the admin application. The new bundle will exist on the
same path that the database requires it on.