Validate all FreeRadius files before publishing
Use FreeRadius in addition to rails/model validations to ensure maximum
safeguarding before pushing a file to S3 and issuing a deployment
command.

This includes:
  1. client.conf
  2. MAB authorize
  3. certificates

If this validation fails, it is treated as an exception, meaning no
error will be surfaced to the user but the engineers will be notified.