:rocket: Deploy Service When Site Tag Changes
This commit updates the spec to test for the case of a user updating a site, and as a result of this the tag changing also.  As a result of this the controllers were also refactored, specifically moving the publish_authorised_clients method to the application controller from the clients Controller.