:coffin: Remove UpcomingExpiringCertificate logic
This commit removes the table UpcomingExpiringCertificate, and the associated logic from the controller. Adding this methodology was going to create a lot of other testing considerations since the UpcomingExpiringCertificate table would have to be kept up to date with the Certificates table. It feels cleaner to lean on the Certificate table to achieve a working solution and allow for refactoring down the line if needed.