✨ Add Expiring Certificate Banner (#562)
This PR adds a banner which appears to the user if there is a certificate uploaded which is going to expire in the next four months. 

![image](https://user-images.githubusercontent.com/26419401/224966476-f5041407-3206-408c-9259-6bc80daacfa0.png)

(I admit to using `style="background-color:#ffdd00"` as I could not figure out how to leverage the [gov-uk yellow](https://design-system.service.gov.uk/styles/colour/) colouring properly. This colour matches that however.)

We elected to not use red since this would indicate an error was occuring.

To do this the application controller now has a new helper method: 
```
helper_method :certificate_expiry_check
...
def certificate_expiry_check
    Certificate.where('expiry_date >= ? AND expiry_date <= ?', Date.today, 4.months.from_now.to_date).exists?
end
```

Since this is in the application controller this can be leveraged in `application.html.erb` and is included here.

```
<% if user_signed_in? and certificate_expiry_check %>
    <%= render "layouts/certificate_banner" %>
<% end %>
```

Testing is encapsulted in `view_certificate_spec.rb`:

```
context "when the user is authenticated" do
...
   context "when a certificate exists but is out of date" do
      let!(:certificate) { create :certificate, expiry_date: Date.today + 2.months }

      it "shows a banner which notifies the user that a certificate is out of date" do
        visit root_path
        expect(page).to have_content "A certificate will expire in less that three months."
      end
  end
end
```

I have tested this locally and load times were not effected.